// apps/backend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  friends: typeof friends;
  http: typeof http;
  sendEmails: typeof sendEmails;
  tasks: typeof tasks;
  users: typeof users;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;